+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/SubEntities.hs
@@ -20,7 +20,7 @@ module Cardano.Ledger.Dijkstra.Rules.SubEntities (
 ) where
 
 import Cardano.Ledger.Address (DirectDeposits (..))
-import Cardano.Ledger.BaseTypes (Globals (networkId), Mismatch (..), Relation (..), ShelleyBase)
+import Cardano.Ledger.BaseTypes
 import Cardano.Ledger.Binary (DecCBOR (..), EncCBOR (..))
 import Cardano.Ledger.Binary.Coders
 import Cardano.Ledger.Coin (Coin)
@@ -28,12 +28,17 @@ import Cardano.Ledger.Conway.Core
 import qualified Cardano.Ledger.Conway.Rules as Conway
 import Cardano.Ledger.Conway.State
 import Cardano.Ledger.Dijkstra.Era (DijkstraEra, SUBCERTS, SUBENTITIES)
+import Cardano.Ledger.Dijkstra.Rules.Entities (
+  EntitiesPredFailure (..),
+ )
 import Cardano.Ledger.Dijkstra.Rules.SubCerts (
   DijkstraSubCertsEvent,
   DijkstraSubCertsPredFailure,
   SubCertsEnv (..),
  )
 import Cardano.Ledger.Dijkstra.TxBody (DijkstraEraTxBody, directDepositsTxBodyL)
+import Cardano.Ledger.Rules.ValidationMode (runTest)
+import qualified Cardano.Ledger.Shelley.Rules as Shelley
 import Control.DeepSeq (NFData)
 import Control.Monad.Trans.Reader (asks)
 import Control.State.Transition.Extended
@@ -41,6 +46,7 @@ import Data.Map.NonEmpty (NonEmptyMap)
 import qualified Data.Map.NonEmpty as NEM
 import qualified Data.Map.Strict as Map
 import Data.Sequence (Seq)
+import Data.Set.NonEmpty (NonEmptySet)
 import GHC.Generics (Generic)
 import Lens.Micro
 
@@ -49,6 +55,11 @@ data SubEntitiesPredFailure era
   | SubWithdrawalsMissingAccounts Withdrawals
   | SubWithdrawalAmountsExceedAccountBalances (NonEmptyMap AccountAddress (Mismatch RelLTEQ Coin))
   | SubDirectDepositsToMissingAccounts DirectDeposits
+  | SubWrongNetworkWithdrawal
+      -- | Expected network id
+      Network
+      -- | Withdrawal accounts with wrong network id
+      (NonEmptySet AccountAddress)
   deriving (Generic)
 
 deriving stock instance
@@ -73,6 +84,7 @@ instance
       SubWithdrawalsMissingAccounts x -> Sum (SubWithdrawalsMissingAccounts @era) 1 !> To x
       SubWithdrawalAmountsExceedAccountBalances x -> Sum (SubWithdrawalAmountsExceedAccountBalances @era) 2 !> To x
       SubDirectDepositsToMissingAccounts x -> Sum (SubDirectDepositsToMissingAccounts @era) 3 !> To x
+      SubWrongNetworkWithdrawal expected wrongs -> Sum (SubWrongNetworkWithdrawal @era) 4 !> To expected !> To wrongs
 
 instance
   ( Era era
@@ -85,6 +97,7 @@ instance
     1 -> SumD SubWithdrawalsMissingAccounts <! From
     2 -> SumD SubWithdrawalAmountsExceedAccountBalances <! From
     3 -> SumD SubDirectDepositsToMissingAccounts <! From
+    4 -> SumD SubWrongNetworkWithdrawal <! From <! From
     n -> Invalid n
 
 newtype SubEntitiesEvent era = SubCertsEvent (Event (EraRule "SUBCERTS" era))
@@ -109,6 +122,12 @@ instance InjectRuleFailure "SUBENTITIES" Conway.ConwayCertsPredFailure DijkstraE
 instance InjectRuleFailure "SUBENTITIES" Conway.ConwayLedgerPredFailure DijkstraEra where
   injectFailure = conwayToDijkstraSubEntitiesPredFailure
 
+instance InjectRuleFailure "SUBENTITIES" EntitiesPredFailure DijkstraEra where
+  injectFailure = entitiesToSubEntitiesPredFailure
+
+instance InjectRuleFailure "SUBENTITIES" Shelley.ShelleyUtxoPredFailure DijkstraEra where
+  injectFailure = injectFailure @"SUBENTITIES" @EntitiesPredFailure . injectFailure @"ENTITIES"
+
 instance
   ( EraTx era
   , DijkstraEraTxBody era
@@ -119,6 +138,7 @@ instance
   , Environment (EraRule "SUBCERTS" era) ~ SubCertsEnv era
   , EraRule "SUBENTITIES" era ~ SUBENTITIES era
   , InjectRuleFailure "SUBENTITIES" SubEntitiesPredFailure era
+  , InjectRuleFailure "SUBENTITIES" Shelley.ShelleyUtxoPredFailure era
   , InjectRuleFailure "SUBENTITIES" Conway.ConwayLedgerPredFailure era
   ) =>
   STS (SUBENTITIES era)
@@ -144,6 +164,7 @@ dijkstraSubEntitiesTransition ::
   , Environment (EraRule "SUBCERTS" era) ~ SubCertsEnv era
   , EraRule "SUBENTITIES" era ~ SUBENTITIES era
   , InjectRuleFailure "SUBENTITIES" SubEntitiesPredFailure era
+  , InjectRuleFailure "SUBENTITIES" Shelley.ShelleyUtxoPredFailure era
   , InjectRuleFailure "SUBENTITIES" Conway.ConwayLedgerPredFailure era
   ) =>
   TransitionRule (SUBENTITIES era)
@@ -156,6 +177,9 @@ dijkstraSubEntitiesTransition = do
       accounts = certState ^. certDStateL . accountsL
 
   network <- liftSTS $ asks networkId
+
+  runTest $ Shelley.validateWrongNetworkWithdrawal network (tx ^. bodyTxL)
+
   let (missingWithdrawals, exceededWithdrawals) =
         case withdrawalsThatExceedAccountBalance withdrawals network accounts of
           Nothing -> (Map.empty, Map.empty)
@@ -191,6 +215,18 @@ conwayToDijkstraSubEntitiesPredFailure = \case
   Conway.ConwayMempoolFailure _ -> impossible "ConwayMempoolFailure"
   Conway.ConwayWithdrawalsMissingAccounts _ -> impossible "ConwayWithdrawalsMissingAccounts"
   Conway.ConwayIncompleteWithdrawals _ -> impossible "ConwayIncompleteWithdrawals"
+  where
+    impossible name = error $ "Impossible: `" <> name <> "` for SUBENTITIES"
+
+entitiesToSubEntitiesPredFailure ::
+  EntitiesPredFailure era -> SubEntitiesPredFailure era
+entitiesToSubEntitiesPredFailure = \case
+  WrongNetworkWithdrawal net addrs -> SubWrongNetworkWithdrawal net addrs
+  CertsFailure _ -> impossible "CertsFailure"
+  WithdrawalsMissingAccounts _ -> impossible "WithdrawalsMissingAccounts"
+  IncompleteWithdrawals _ -> impossible "IncompleteWithdrawals"
+  WithdrawalAmountsExceedAccountBalances _ -> impossible "WithdrawalAmountsExceedAccountBalances"
+  DirectDepositsToMissingAccounts _ -> impossible "DirectDepositsToMissingAccounts"
   where
     impossible name = error $ "Impossible: `" <> name <> "` for SUBENTITIES"