+++ eras/dijkstra/impl/src/Cardano/Ledger/Dijkstra/Rules/SubUtxo.hs
@@ -85,11 +85,6 @@ data DijkstraSubUtxoPredFailure era
       Network
       -- | the set of addresses with incorrect network IDs
       (NonEmptySet Addr)
-  | SubWrongNetworkWithdrawal
-      -- | the expected network id
-      Network
-      -- | the set of reward addresses with incorrect network IDs
-      (NonEmptySet AccountAddress)
   | -- | list of supplied bad transaction outputs
     SubOutputBootAddrAttrsTooBig (NonEmpty (TxOut era))
   | -- | list of supplied bad transaction output triples (actualSize,PParameterMaxValue,TxOut)
@@ -257,7 +252,6 @@ dijkstraSubUtxoTransition = do
 
   netId <- liftSTS $ asks networkId
   runTestOnSignal $ Shelley.validateWrongNetwork netId allOutputs
-  runTestOnSignal $ Shelley.validateWrongNetworkWithdrawal netId txBody
   runTestOnSignal $ validateWrongNetworkInDirectDeposit netId txBody
   runTestOnSignal $ Alonzo.validateWrongNetworkInTxBody netId txBody
 
@@ -283,7 +277,6 @@ instance
       SubMaxTxSizeUTxO mm -> Sum SubMaxTxSizeUTxO 2 !> To mm
       SubInputSetEmptyUTxO -> Sum SubInputSetEmptyUTxO 3
       SubWrongNetwork right wrongs -> Sum (SubWrongNetwork @era) 4 !> To right !> To wrongs
-      SubWrongNetworkWithdrawal right wrongs -> Sum (SubWrongNetworkWithdrawal @era) 5 !> To right !> To wrongs
       SubOutputBootAddrAttrsTooBig outs -> Sum (SubOutputBootAddrAttrsTooBig @era) 6 !> To outs
       SubOutputTooBigUTxO outs -> Sum (SubOutputTooBigUTxO @era) 7 !> To outs
       SubWrongNetworkInTxBody mm -> Sum SubWrongNetworkInTxBody 8 !> To mm
@@ -305,7 +298,6 @@ instance
     2 -> SumD SubMaxTxSizeUTxO <! From
     3 -> SumD SubInputSetEmptyUTxO
     4 -> SumD SubWrongNetwork <! From <! From
-    5 -> SumD SubWrongNetworkWithdrawal <! From <! From
     6 -> SumD SubOutputBootAddrAttrsTooBig <! From
     7 -> SumD SubOutputTooBigUTxO <! From
     8 -> SumD SubWrongNetworkInTxBody <! From
@@ -325,7 +317,6 @@ dijkstraUtxoToDijkstraSubUtxoPredFailure = \case
   FeeTooSmallUTxO _ -> error "Impossible: `FeeTooSmallUTxO` for SUBUTXO"
   ValueNotConservedUTxO _ -> error "Impossible: `ValueNotConservedUTxO` for SUBUTXO"
   WrongNetwork x y -> SubWrongNetwork x y
-  WrongNetworkWithdrawal x y -> SubWrongNetworkWithdrawal x y
   OutputBootAddrAttrsTooBig xs -> SubOutputBootAddrAttrsTooBig xs
   OutputTooBigUTxO xs -> SubOutputTooBigUTxO xs
   InsufficientCollateral _ _ -> error "Impossible: `InsufficientCollateral` for SUBUTXO"